requirements.txt
aiohttp
scikit-learn
numpy
requests
pymongo
redis
flask
python-dotenv
twitchio
gunicorn